Dep inject deserializer instead of global. Extract types
